requirements.txt
python-telegram-bot>=20.0,<21.0
transmission-rpc>=4.0.0
requests>=2.0.0
bencodepy>=0.9.5
pytest>=7.0.0

